+++ haskell-src/Concordium/GRPC2.hs
@@ -883,6 +883,12 @@ instance ToProto Parameters.FinalizationCommitteeParameters where
         ProtoFields.maximumFinalizers .= _fcpMaxFinalizers
         ProtoFields.finalizerRelativeStakeThreshold .= toProto _fcpFinalizerRelativeStakeThreshold
 
+instance ToProto Parameters.ValidatorScoreParameters where
+    type Output Parameters.ValidatorScoreParameters = Proto.ValidatorScoreParameters
+    toProto Parameters.ValidatorScoreParameters{..} =
+        Proto.make $
+            ProtoFields.maximumMissedRounds .= _vspMaxMissedRounds
+
 instance ToProto (Parameters.ConsensusParameters' 'Parameters.ConsensusParametersVersion1) where
     type Output (Parameters.ConsensusParameters' 'Parameters.ConsensusParametersVersion1) = Proto.ConsensusParametersV1
     toProto Parameters.ConsensusParametersV1{..} = Proto.make $ do
@@ -923,6 +929,7 @@ convertUpdatePayload ut pl = case (ut, pl) of
     (Updates.UpdateMinBlockTime, Updates.MinBlockTimeUpdatePayload mbt) -> Right . Proto.make $ ProtoFields.minBlockTimeUpdate .= toProto mbt
     (Updates.UpdateBlockEnergyLimit, Updates.BlockEnergyLimitUpdatePayload bel) -> Right . Proto.make $ ProtoFields.blockEnergyLimitUpdate .= toProto bel
     (Updates.UpdateFinalizationCommitteeParameters, Updates.FinalizationCommitteeParametersUpdatePayload fcp) -> Right . Proto.make $ ProtoFields.finalizationCommitteeParametersUpdate .= toProto fcp
+    (Updates.UpdateValidatorScoreParameters, Updates.ValidatorScoreParametersUpdatePayload vsp) -> Right . Proto.make $ ProtoFields.validatorScoreParametersUpdate .= toProto vsp
     _ -> Left CEInvalidUpdateResult
 
 -- | The different conversions errors possible in @toBlockItemStatus@ (and the helper to* functions it calls).
@@ -1583,6 +1590,7 @@ instance ToProto Updates.UpdateType where
     toProto Updates.UpdateMinBlockTime = Proto.UPDATE_MIN_BLOCK_TIME
     toProto Updates.UpdateBlockEnergyLimit = Proto.UPDATE_BLOCK_ENERGY_LIMIT
     toProto Updates.UpdateFinalizationCommitteeParameters = Proto.UPDATE_FINALIZATION_COMMITTEE_PARAMETERS
+    toProto Updates.UpdateValidatorScoreParameters = Proto.UPDATE_VALIDATOR_SCORE_PARAMETERS
 
 instance ToProto TransactionType where
     type Output TransactionType = Proto.TransactionType
@@ -1969,6 +1977,14 @@ instance ToProto TxTypes.SpecialTransactionOutcome where
                         ProtoFields.bakerReward .= toProto stoBakerReward
                         ProtoFields.finalizationReward .= toProto stoFinalizationReward
                     )
+    toProto TxTypes.ValidatorPrimedForSuspension{..} =
+        Proto.make $
+            ProtoFields.validatorSuspended
+                .= Proto.make (ProtoFields.bakerId .= toProto vpfsBakerId)
+    toProto TxTypes.ValidatorSuspended{..} =
+        Proto.make $
+            ProtoFields.validatorSuspended
+                .= Proto.make (ProtoFields.bakerId .= toProto vsBakerId)
 
 instance ToProto (TransactionTime, QueryTypes.PendingUpdateEffect) where
     type Output (TransactionTime, QueryTypes.PendingUpdateEffect) = Proto.PendingUpdate
@@ -1999,6 +2015,7 @@ instance ToProto (TransactionTime, QueryTypes.PendingUpdateEffect) where
             QueryTypes.PUEMinBlockTime minBlockTime -> ProtoFields.minBlockTime .= toProto minBlockTime
             QueryTypes.PUEBlockEnergyLimit blockEnergyLimit -> ProtoFields.blockEnergyLimit .= toProto blockEnergyLimit
             QueryTypes.PUEFinalizationCommitteeParameters finalizationCommitteeParameters -> ProtoFields.finalizationCommitteeParameters .= toProto finalizationCommitteeParameters
+            QueryTypes.PUEValidatorScoreParameters validatorScoreParameters -> ProtoFields.validatorScoreParameters .= toProto validatorScoreParameters
 
 instance ToProto QueryTypes.NextUpdateSequenceNumbers where
     type Output QueryTypes.NextUpdateSequenceNumbers = Proto.NextUpdateSequenceNumbers
@@ -2023,6 +2040,7 @@ instance ToProto QueryTypes.NextUpdateSequenceNumbers where
         ProtoFields.minBlockTime .= toProto _nusnMinBlockTime
         ProtoFields.blockEnergyLimit .= toProto _nusnBlockEnergyLimit
         ProtoFields.finalizationCommitteeParameters .= toProto _nusnFinalizationCommitteeParameters
+        ProtoFields.validatorScoreParameters .= toProto _nusnValidatorScoreParameters
 
 instance ToProto Epoch where
     type Output Epoch = Proto.Epoch