+++ haskell-src/Concordium/Types/Parameters.hs
@@ -329,6 +329,8 @@ module Concordium.Types.Parameters (
     -- | Parameters for baker pools. Prior to P4, this is just the minimum stake threshold
     --  for becoming a baker.
     cpPoolParameters,
+    -- | Parameters for validator suspension. Available since P8.
+    cpValidatorScoreParameters,
     EChainParameters (..),
     ChainParameters,
     putChainParameters,
@@ -357,6 +359,10 @@ module Concordium.Types.Parameters (
     -- | 'FinalizationCommitteeParameters' wrapped in an 'OParam'
     --  supporting ''PTFinalizationCommitteeParameters'.
     OFinalizationCommitteeParameters,
+
+    -- * Validator score parameters
+    ValidatorScoreParameters (..),
+    vspMaxMissedRounds,
 
     -- * Authorizations version
 
@@ -412,6 +418,7 @@ module Concordium.Types.Parameters (
     PTCooldownParametersAccessStructureSym0,
     PTFinalizationProofSym0,
     PTFinalizationCommitteeParametersSym0,
+    PTValidatorScoreParametersSym0,
 ) where
 
 import Control.Monad
@@ -621,6 +628,8 @@ $( singletons
               PTFinalizationProof
             | -- Finalization committee selection for V2 consensus
               PTFinalizationCommitteeParameters
+            | -- Maximal score a validator can reach before it gets suspended
+              PTValidatorScoreParameters
 
         -- \|Whether a particular parameter is supported at a particular 'ChainParametersVersion'.
         isSupported :: ParameterType -> ChainParametersVersion -> Bool
@@ -647,6 +656,10 @@ $( singletons
         isSupported PTFinalizationCommitteeParameters ChainParametersV1 = False
         isSupported PTFinalizationCommitteeParameters ChainParametersV2 = True
         isSupported PTFinalizationCommitteeParameters ChainParametersV3 = True
+        isSupported PTValidatorScoreParameters ChainParametersV0 = False
+        isSupported PTValidatorScoreParameters ChainParametersV1 = False
+        isSupported PTValidatorScoreParameters ChainParametersV2 = False
+        isSupported PTValidatorScoreParameters ChainParametersV3 = True
         |]
  )
 
@@ -1639,6 +1652,40 @@ type OFinalizationCommitteeParameters (pv :: ProtocolVersion) =
         (ChainParametersVersionFor pv)
         FinalizationCommitteeParameters
 
+-- * Validator score parameters
+
+-- | Score specific parameters.
+newtype ValidatorScoreParameters = ValidatorScoreParameters
+    { -- | Maximal number of rounds a validator can miss before it gets suspended.
+      _vspMaxMissedRounds :: Word64
+    }
+    deriving (Eq, Show)
+
+makeLenses ''ValidatorScoreParameters
+
+instance HashableTo Hash.Hash ValidatorScoreParameters where
+    getHash = Hash.hash . encode
+
+instance (Monad m) => MHashableTo m Hash.Hash ValidatorScoreParameters
+
+instance Serialize ValidatorScoreParameters where
+    put ValidatorScoreParameters{..} = do
+        put _vspMaxMissedRounds
+    get = do
+        _vspMaxMissedRounds <- get
+        return ValidatorScoreParameters{..}
+
+instance ToJSON ValidatorScoreParameters where
+    toJSON ValidatorScoreParameters{..} =
+        object
+            [ "maximumMissedRounds" AE..= _vspMaxMissedRounds
+            ]
+
+instance FromJSON ValidatorScoreParameters where
+    parseJSON = withObject "ValidatorScoreParameters" $ \o -> do
+        _vspMaxMissedRounds <- o .: "maximumMissedRounds"
+        return ValidatorScoreParameters{..}
+
 -- * Consensus parameters
 
 -- | Constraint on a type level 'ConsensusParametersVersion' that can be used to get a corresponding
@@ -1771,7 +1818,10 @@ data ChainParameters' (cpv :: ChainParametersVersion) = ChainParameters
       -- | The finalization committee parameters.
       --  These parameters are introduced as part of protocol 6 (cpv2).
       --  The set of parameters shares the 'Authorization' with the '_cpPoolParameters'.
-      _cpFinalizationCommitteeParameters :: !(OParam 'PTFinalizationCommitteeParameters cpv FinalizationCommitteeParameters)
+      _cpFinalizationCommitteeParameters :: !(OParam 'PTFinalizationCommitteeParameters cpv FinalizationCommitteeParameters),
+      -- | The score parameters.
+      --  These parameters are introduced as part of protocol 8 (cpv3).
+      _cpValidatorScoreParameters :: !(OParam 'PTValidatorScoreParameters cpv ValidatorScoreParameters)
     }
     deriving (Eq, Show)
 
@@ -1803,6 +1853,7 @@ putChainParameters ChainParameters{..} = do
     put _cpFoundationAccount
     putPoolParameters _cpPoolParameters
     put _cpFinalizationCommitteeParameters
+    put _cpValidatorScoreParameters
 
 -- | Deserialize a 'ChainParameters''.
 getChainParameters :: forall cpv. (IsChainParametersVersion cpv) => Get (ChainParameters' cpv)
@@ -1816,6 +1867,7 @@ getChainParameters = do
     _cpFoundationAccount <- get
     _cpPoolParameters <- withIsPoolParametersVersionFor (chainParametersVersion @cpv) get
     _cpFinalizationCommitteeParameters <- get
+    _cpValidatorScoreParameters <- get
     return ChainParameters{..}
 
 instance (IsChainParametersVersion cpv) => Serialize (ChainParameters' cpv) where
@@ -1851,6 +1903,7 @@ parseJSONForCPV0 =
                     .: "minimumThresholdForBaking"
         let _cpTimeParameters = NoParam
             _cpFinalizationCommitteeParameters = NoParam
+            _cpValidatorScoreParameters = NoParam
         return ChainParameters{..}
 
 parseJSONForCPV1 :: Value -> Parser (ChainParameters' 'ChainParametersV1)
@@ -1883,6 +1936,7 @@ parseJSONForCPV1 =
             _ppPassiveCommissions = CommissionRates{..}
             _ppCommissionBounds = CommissionRanges{..}
             _cpFinalizationCommitteeParameters = NoParam
+            _cpValidatorScoreParameters = NoParam
         return ChainParameters{..}
 
 parseJSONForCPV2 :: Value -> Parser (ChainParameters' 'ChainParametersV2)
@@ -1924,11 +1978,12 @@ parseJSONForCPV2 =
             _ppCommissionBounds = CommissionRanges{..}
             _cpFinalizationCommitteeParameters = SomeParam FinalizationCommitteeParameters{..}
             _cpConsensusParameters = ConsensusParametersV1{..}
+            _cpValidatorScoreParameters = NoParam
         return ChainParameters{..}
 
 parseJSONForCPV3 :: Value -> Parser (ChainParameters' 'ChainParametersV3)
 parseJSONForCPV3 =
-    withObject "ChainParametersV2" $ \v -> do
+    withObject "ChainParametersV3" $ \v -> do
         _cpEuroPerEnergy <- v .: "euroPerEnergy"
         _cpMicroGTUPerEuro <- v .: "microGTUPerEuro"
         _cpPoolOwnerCooldown <- v .: "poolOwnerCooldown"
@@ -1965,6 +2020,10 @@ parseJSONForCPV3 =
             _ppCommissionBounds = CommissionRanges{..}
             _cpFinalizationCommitteeParameters = SomeParam FinalizationCommitteeParameters{..}
             _cpConsensusParameters = ConsensusParametersV1{..}
+
+        _vspMaxMissedRounds <- v .: "maximumMissedRounds"
+        let _cpValidatorScoreParameters = SomeParam ValidatorScoreParameters{..}
+
         return ChainParameters{..}
 
 instance forall cpv. (IsChainParametersVersion cpv) => FromJSON (ChainParameters' cpv) where