+++ rust-src/concordium_base/src/updates.rs
@@ -627,6 +627,14 @@ pub struct FinalizationCommitteeParameters {
     pub finalizers_relative_stake_threshold: PartsPerHundredThousands,
 }
 
+#[derive(Debug, common::Serialize, Clone, Copy, SerdeSerialize, SerdeDeserialize)]
+/// Validator score parameters. These parameters control the threshold of
+/// maximal missed rounds before a validator gets suspended.
+pub struct ValidatorScoreParameters {
+    /// Maximal number of missed rounds before a validator gets suspended.
+    pub max_missed_rounds: u64,
+}
+
 #[derive(SerdeSerialize, SerdeDeserialize, Debug, Clone)]
 #[serde(tag = "updateType", content = "update")]
 /// The type of an update payload.
@@ -675,6 +683,8 @@ pub enum UpdatePayload {
     BlockEnergyLimitCPV2(Energy),
     #[serde(rename = "finalizationCommitteeParametersCPV2")]
     FinalizationCommitteeParametersCPV2(FinalizationCommitteeParameters),
+    #[serde(rename = "validatorScoreParametersCPV3")]
+    ValidatorScoreParametersCPV3(ValidatorScoreParameters),
 }
 
 #[derive(SerdeSerialize, SerdeDeserialize, Debug, Clone, Copy)]
@@ -735,6 +745,9 @@ pub enum UpdateType {
     /// Update of the finalization committee parameters. Only applies to
     /// protocol version [`P6`](ProtocolVersion::P6) and up.
     UpdateFinalizationCommitteeParameters,
+    /// Update of the validator score parameters. Only applies to
+    /// protocol version [`P8`](ProtocolVersion::P8) and up.
+    UpdateValidatorScoreParameters,
 }
 
 impl UpdatePayload {
@@ -765,6 +778,7 @@ impl UpdatePayload {
             UpdatePayload::FinalizationCommitteeParametersCPV2(_) => {
                 UpdateFinalizationCommitteeParameters
             }
+            UpdatePayload::ValidatorScoreParametersCPV3(_) => UpdateValidatorScoreParameters,
         }
     }
 }
@@ -982,6 +996,10 @@ impl Serial for UpdatePayload {
                 22u8.serial(out);
                 update.serial(out)
             }
+            UpdatePayload::ValidatorScoreParametersCPV3(update) => {
+                23u8.serial(out);
+                update.serial(out)
+            }
         }
     }
 }
@@ -1035,6 +1053,7 @@ impl Deserial for UpdatePayload {
             22u8 => Ok(UpdatePayload::FinalizationCommitteeParametersCPV2(
                 source.get()?,
             )),
+            23u8 => Ok(UpdatePayload::ValidatorScoreParametersCPV3(source.get()?)),
             tag => anyhow::bail!("Unknown update payload tag {}", tag),
         }
     }